+++ packages/image_picker/image_picker_platform_interface/test/new_method_channel_image_picker_test.dart
@@ -18,6 +18,7 @@ void main() {
     dynamic returnValue = '';
 
     setUp(() {
+      returnValue = '';
       picker.channel.setMockMethodCallHandler((MethodCall methodCall) async {
         log.add(methodCall);
         return returnValue;
@@ -137,6 +138,29 @@ void main() {
               'cameraDevice': 0
             }),
           ],
+        );
+      });
+
+      test('does not accept a invalid imageQuality argument', () {
+        expect(
+          () => picker.pickImage(imageQuality: -1, source: ImageSource.gallery),
+          throwsArgumentError,
+        );
+
+        expect(
+          () =>
+              picker.pickImage(imageQuality: 101, source: ImageSource.gallery),
+          throwsArgumentError,
+        );
+
+        expect(
+          () => picker.pickImage(imageQuality: -1, source: ImageSource.camera),
+          throwsArgumentError,
+        );
+
+        expect(
+          () => picker.pickImage(imageQuality: 101, source: ImageSource.camera),
+          throwsArgumentError,
         );
       });
 
@@ -292,6 +316,19 @@ void main() {
 
         expect(
           () => picker.pickMultiImage(maxHeight: -1.0),
+          throwsArgumentError,
+        );
+      });
+
+      test('does not accept a invalid imageQuality argument', () {
+        returnValue = ['0', '1'];
+        expect(
+          () => picker.pickMultiImage(imageQuality: -1),
+          throwsArgumentError,
+        );
+
+        expect(
+          () => picker.pickMultiImage(imageQuality: 101),
           throwsArgumentError,
         );
       });