+++ script/tool/lib/src/common.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io' as io;
 import 'dart:math';
 
@@ -15,7 +16,9 @@ import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 import 'package:yaml/yaml.dart';
 
-typedef void Print(Object object);
+/// The signature for a print handler for commands that allow overriding the
+/// print destination.
+typedef Print = void Function(Object object);
 
 /// Key for windows platform.
 const String kWindows = 'windows';
@@ -53,8 +56,9 @@ bool isFlutterPackage(FileSystemEntity entity, FileSystem fileSystem) {
   try {
     final File pubspecFile =
         fileSystem.file(p.join(entity.path, 'pubspec.yaml'));
-    final YamlMap pubspecYaml = loadYaml(pubspecFile.readAsStringSync());
-    final YamlMap dependencies = pubspecYaml['dependencies'];
+    final YamlMap pubspecYaml =
+        loadYaml(pubspecFile.readAsStringSync()) as YamlMap;
+    final YamlMap dependencies = pubspecYaml['dependencies'] as YamlMap;
     if (dependencies == null) {
       return false;
     }
@@ -89,16 +93,17 @@ bool pluginSupportsPlatform(
   try {
     final File pubspecFile =
         fileSystem.file(p.join(entity.path, 'pubspec.yaml'));
-    final YamlMap pubspecYaml = loadYaml(pubspecFile.readAsStringSync());
-    final YamlMap flutterSection = pubspecYaml['flutter'];
+    final YamlMap pubspecYaml =
+        loadYaml(pubspecFile.readAsStringSync()) as YamlMap;
+    final YamlMap flutterSection = pubspecYaml['flutter'] as YamlMap;
     if (flutterSection == null) {
       return false;
     }
-    final YamlMap pluginSection = flutterSection['plugin'];
+    final YamlMap pluginSection = flutterSection['plugin'] as YamlMap;
     if (pluginSection == null) {
       return false;
     }
-    final YamlMap platforms = pluginSection['platforms'];
+    final YamlMap platforms = pluginSection['platforms'] as YamlMap;
     if (platforms == null) {
       // Legacy plugin specs are assumed to support iOS and Android.
       if (!pluginSection.containsKey('platforms')) {
@@ -153,12 +158,16 @@ void printErrorAndExit({@required String errorMessage, int exitCode = 1}) {
 
 /// Error thrown when a command needs to exit with a non-zero exit code.
 class ToolExit extends Error {
+  /// Creates a tool exit with the given [exitCode].
   ToolExit(this.exitCode);
 
+  /// The code that the process should exit with.
   final int exitCode;
 }
 
-abstract class PluginCommand extends Command<Null> {
+/// Interface definition for all commands in this tool.
+abstract class PluginCommand extends Command<void> {
+  /// Creates a command to operate on [packagesDir] with the given environment.
   PluginCommand(
     this.packagesDir,
     this.fileSystem, {
@@ -230,23 +239,25 @@ abstract class PluginCommand extends Command<Null> {
   int _shardIndex;
   int _shardCount;
 
+  /// The shard of the overall command execution that this instance should run.
   int get shardIndex {
     if (_shardIndex == null) {
-      checkSharding();
+      _checkSharding();
     }
     return _shardIndex;
   }
 
+  /// The number of shards this command is divided into.
   int get shardCount {
     if (_shardCount == null) {
-      checkSharding();
+      _checkSharding();
     }
     return _shardCount;
   }
 
-  void checkSharding() {
-    final int shardIndex = int.tryParse(argResults[_shardIndexArg]);
-    final int shardCount = int.tryParse(argResults[_shardCountArg]);
+  void _checkSharding() {
+    final int shardIndex = int.tryParse(argResults[_shardIndexArg] as String);
+    final int shardCount = int.tryParse(argResults[_shardCountArg] as String);
     if (shardIndex == null) {
       usageException('$_shardIndexArg must be an integer');
     }
@@ -281,7 +292,7 @@ abstract class PluginCommand extends Command<Null> {
     final int start = min(shardIndex * shardSize, allPlugins.length);
     final int end = min(start + shardSize, allPlugins.length);
 
-    for (Directory plugin in allPlugins.sublist(start, end)) {
+    for (final Directory plugin in allPlugins.sublist(start, end)) {
       yield plugin;
     }
   }
@@ -301,25 +312,28 @@ abstract class PluginCommand extends Command<Null> {
   ///    "client library" package, which declares the API for the plugin, as
   ///    well as one or more platform-specific implementations.
   Stream<Directory> _getAllPlugins() async* {
-    Set<String> plugins = Set<String>.from(argResults[_pluginsArg]);
+    Set<String> plugins =
+        Set<String>.from(argResults[_pluginsArg] as List<String>);
     final Set<String> excludedPlugins =
-        Set<String>.from(argResults[_excludeArg]);
-    final bool runOnChangedPackages = argResults[_runOnChangedPackagesArg];
+        Set<String>.from(argResults[_excludeArg] as List<String>);
+    final bool runOnChangedPackages =
+        argResults[_runOnChangedPackagesArg] as bool;
     if (plugins.isEmpty && runOnChangedPackages) {
       plugins = await _getChangedPackages();
     }
 
-    await for (FileSystemEntity entity
+    await for (final FileSystemEntity entity
         in packagesDir.list(followLinks: false)) {
       // A top-level Dart package is a plugin package.
       if (_isDartPackage(entity)) {
         if (!excludedPlugins.contains(entity.basename) &&
             (plugins.isEmpty || plugins.contains(p.basename(entity.path)))) {
-          yield entity;
+          yield entity as Directory;
         }
       } else if (entity is Directory) {
         // Look for Dart packages under this top-level directory.
-        await for (FileSystemEntity subdir in entity.list(followLinks: false)) {
+        await for (final FileSystemEntity subdir
+            in entity.list(followLinks: false)) {
           if (_isDartPackage(subdir)) {
             // If --plugin=my_plugin is passed, then match all federated
             // plugins under 'my_plugin'. Also match if the exact plugin is
@@ -334,7 +348,7 @@ abstract class PluginCommand extends Command<Null> {
                 (plugins.isEmpty ||
                     plugins.contains(relativePath) ||
                     plugins.contains(basenamePath))) {
-              yield subdir;
+              yield subdir as Directory;
             }
           }
         }
@@ -350,7 +364,7 @@ abstract class PluginCommand extends Command<Null> {
   /// Returns all Dart package folders (typically, plugin + example) of the
   /// plugins involved in this command execution.
   Stream<Directory> getPackages() async* {
-    await for (Directory plugin in getPlugins()) {
+    await for (final Directory plugin in getPlugins()) {
       yield plugin;
       yield* plugin
           .list(recursive: true, followLinks: false)
@@ -401,7 +415,7 @@ abstract class PluginCommand extends Command<Null> {
   /// Throws tool exit if [gitDir] nor root directory is a git directory.
   Future<GitVersionFinder> retrieveVersionFinder() async {
     final String rootDir = packagesDir.parent.absolute.path;
-    String baseSha = argResults[_kBaseSha];
+    final String baseSha = argResults[_kBaseSha] as String;
 
     GitDir baseGitDir = gitDir;
     if (baseGitDir == null) {
@@ -424,14 +438,14 @@ abstract class PluginCommand extends Command<Null> {
     final List<String> allChangedFiles =
         await gitVersionFinder.getChangedFiles();
     final Set<String> packages = <String>{};
-    allChangedFiles.forEach((String path) {
+    for (final String path in allChangedFiles) {
       final List<String> pathComponents = path.split('/');
       final int packagesIndex =
           pathComponents.indexWhere((String element) => element == 'packages');
       if (packagesIndex != -1) {
         packages.add(pathComponents[packagesIndex + 1]);
       }
-    });
+    }
     if (packages.isNotEmpty) {
       final String changedPackages = packages.join(',');
       print(changedPackages);
@@ -446,6 +460,7 @@ abstract class PluginCommand extends Command<Null> {
 /// We use this instead of directly running the process so it can be overridden
 /// in tests.
 class ProcessRunner {
+  /// Creates a new process runner.
   const ProcessRunner();
 
   /// Run the [executable] with [args] and stream output to stderr and stdout.
@@ -490,8 +505,8 @@ class ProcessRunner {
   Future<io.ProcessResult> run(String executable, List<String> args,
       {Directory workingDir,
       bool exitOnError = false,
-      stdoutEncoding = io.systemEncoding,
-      stderrEncoding = io.systemEncoding}) async {
+      Encoding stdoutEncoding = io.systemEncoding,
+      Encoding stderrEncoding = io.systemEncoding}) async {
     return io.Process.run(executable, args,
         workingDirectory: workingDir?.path,
         stdoutEncoding: stdoutEncoding,
@@ -569,24 +584,32 @@ class GitVersionFinder {
   Future<List<String>> getChangedFiles() async {
     final String baseSha = await _getBaseSha();
     final io.ProcessResult changedFilesCommand = await baseGitDir
-        .runCommand(<String>['diff', '--name-only', '$baseSha', 'HEAD']);
+        .runCommand(<String>['diff', '--name-only', baseSha, 'HEAD']);
     print('Determine diff with base sha: $baseSha');
-    final String changedFilesStdout = changedFilesCommand.stdout.toString()  ?? '';
+    final String changedFilesStdout =
+        changedFilesCommand.stdout.toString() ?? '';
     if (changedFilesStdout.isEmpty) {
       return <String>[];
     }
-    final List<String> changedFiles = changedFilesStdout
-        .split('\n')
-          ..removeWhere((element) => element.isEmpty);
+    final List<String> changedFiles = changedFilesStdout.split('\n')
+      ..removeWhere((String element) => element.isEmpty);
     return changedFiles.toList();
   }
 
-  /// Get the package version specified in the pubspec file in `pubspecPath` and at the revision of `gitRef`.
-  Future<Version> getPackageVersion(String pubspecPath, String gitRef) async {
-    final io.ProcessResult gitShow =
-        await baseGitDir.runCommand(<String>['show', '$gitRef:$pubspecPath']);
-    final String fileContent = gitShow.stdout;
-    final String versionString = loadYaml(fileContent)['version'];
+  /// Get the package version specified in the pubspec file in `pubspecPath` and
+  /// at the revision of `gitRef` (defaulting to the base if not provided).
+  Future<Version> getPackageVersion(String pubspecPath, {String gitRef}) async {
+    final String ref = gitRef ?? (await _getBaseSha());
+
+    io.ProcessResult gitShow;
+    try {
+      gitShow =
+          await baseGitDir.runCommand(<String>['show', '$ref:$pubspecPath']);
+    } on io.ProcessException {
+      return null;
+    }
+    final String fileContent = gitShow.stdout as String;
+    final String versionString = loadYaml(fileContent)['version'] as String;
     return versionString == null ? null : Version.parse(versionString);
   }